+++ lib/api/src/sys/module.rs
@@ -5,8 +5,8 @@ use std::io;
 use std::path::Path;
 use std::sync::Arc;
 use thiserror::Error;
-use wasmer_compiler::ArtifactCreate;
-use wasmer_compiler::UniversalArtifact;
+use wasmer_compiler::Artifact;
+use wasmer_compiler::ArtifactMetadata;
 #[cfg(feature = "wat")]
 use wasmer_types::WasmError;
 use wasmer_types::{
@@ -51,7 +51,7 @@ pub struct Module {
     //
     // In the future, this code should be refactored to properly describe the
     // ownership of the code and its metadata.
-    artifact: Arc<UniversalArtifact>,
+    artifact: Arc<Artifact>,
     module_info: Arc<ModuleInfo>,
     store: Store,
 }
@@ -271,7 +271,7 @@ impl Module {
         Ok(Self::from_artifact(store, artifact))
     }
 
-    fn from_artifact(store: &Store, artifact: Arc<UniversalArtifact>) -> Self {
+    fn from_artifact(store: &Store, artifact: Arc<Artifact>) -> Self {
         Self {
             store: store.clone(),
             module_info: Arc::new(artifact.create_module_info()),
@@ -449,7 +449,7 @@ impl Module {
     /// this functionality is required for some core functionality though, like
     /// the object file engine.
     #[doc(hidden)]
-    pub fn artifact(&self) -> &Arc<UniversalArtifact> {
+    pub fn artifact(&self) -> &Arc<Artifact> {
         &self.artifact
     }
 }