+++ lib/compiler/src/engine/mod.rs
@@ -6,7 +6,7 @@ mod trap;
 mod tunables;
 
 #[cfg(feature = "translator")]
-mod universal;
+mod artifact;
 
 pub use self::error::{InstantiationError, LinkError};
 pub use self::resolver::resolve_imports;
@@ -14,4 +14,4 @@ pub use self::trap::*;
 pub use self::tunables::Tunables;
 
 #[cfg(feature = "translator")]
-pub use self::universal::*;
+pub use self::artifact::*;