+++ lib/compiler/src/lib.rs
@@ -51,19 +51,19 @@ mod lib {
     }
 }
 
-mod artifact;
 #[cfg(not(target_arch = "wasm32"))]
 mod engine;
+mod traits;
 
-pub use crate::artifact::*;
 #[cfg(not(target_arch = "wasm32"))]
 pub use crate::engine::*;
+pub use crate::traits::*;
 
 #[cfg(feature = "translator")]
-mod universal_artifact;
+mod build;
 
 #[cfg(feature = "translator")]
-pub use self::universal_artifact::*;
+pub use self::build::*;
 
 #[cfg(feature = "translator")]
 mod compiler;